refactor: move config to core
