+++ src/core/config/config-base.jsonc
